requirements.txt
# Core system monitoring dependencies
psutil>=5.8.0
py-cpuinfo>=8.0.0

# GPU monitoring (NVIDIA)
pynvml>=11.4.1

# PyTorch for CUDA detection (optional but recommended)
torch>=1.9.0

# Web framework dependencies
aiohttp>=3.8.0
aiohttp-cors>=0.7.0

# Desktop integration (Windows)
winshell>=0.6
pywin32>=306

# System tray interface (optional)
pystray>=0.19.0
pillow>=10.0.0
